Implement command parser
