+++ src/main/java/john/Command.java
@@ -1,0 +1,5 @@
+package john;
+
+public enum Command {
+    MARK, UNMARK, LIST, TODO, DEADLINE, EVENT, DELETE, BYE, WRONG;
+}
